test(list): migrate inset
