+++ core/src/components/list/test/inset/list.e2e.ts-snapshots/list-inset-diff-ios-rtl-Mobile-Safari-linux.png
