+++ core/src/components/list/test/inset/list.e2e.ts-snapshots/list-inset-inset-lines-ios-ltr-Mobile-Safari-linux.png
