+++ core/src/components/list/test/inset/list.e2e.ts-snapshots/list-inset-inset-lines-md-ltr-Mobile-Firefox-linux.png
