+++ core/src/components/list/test/inset/list.e2e.ts-snapshots/list-inset-no-lines-ios-ltr-Mobile-Firefox-linux.png
